+++ src/components/TimeSeriesCharts.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react';
-import {createStyles, withStyles, WithStyles} from '@material-ui/core/styles';
-import {Theme} from '@material-ui/core';
+import { createStyles, withStyles, WithStyles } from '@material-ui/core/styles';
+import { Theme } from '@material-ui/core';
 
-import {WithLocale} from '../util/lang';
+import { WithLocale } from '../util/lang';
 import TimeSeriesChart from './TimeSeriesChart';
-import {Time, TimeRange, TimeSeriesGroup} from '../model/timeSeries';
-import {PlaceInfo} from "../model/place";
+import { Time, TimeRange, TimeSeriesGroup } from '../model/timeSeries';
+import { Place, PlaceInfo } from "../model/place";
 import TimeRangeSlider from './TimeRangeSlider';
 
 
@@ -43,6 +43,8 @@ interface TimeSeriesChartsProps extends WithStyles<typeof styles>, WithLocale {
 
     removeTimeSeriesGroup?: (id: string) => void;
     placeInfos?: { [placeId: string]: PlaceInfo };
+    selectPlace: (placeId: string | null, places: Place[], showInMap: boolean) => void;
+    places: Place[];
 }
 
 class TimeSeriesCharts extends React.Component<TimeSeriesChartsProps> {
@@ -54,37 +56,39 @@ class TimeSeriesCharts extends React.Component<TimeSeriesChartsProps> {
             selectedTime, selectedTimeRange,
             dataTimeRange, selectTime, selectTimeRange,
             removeTimeSeriesGroup, showPointsOnly, showErrorBars,
-            placeInfos
+            placeInfos, places, selectPlace
         } = this.props;
 
         const charts = timeSeriesGroups.map((timeSeriesGroup: TimeSeriesGroup) => {
-                const completed = timeSeriesGroup.timeSeriesArray.map(item => (
-                        item.dataProgress ? 100 * item.dataProgress : 0
-                    )
-                );
+            const completed = timeSeriesGroup.timeSeriesArray.map(item => (
+                item.dataProgress ? 100 * item.dataProgress : 0
+            ));
 
-                return (
-                    <TimeSeriesChart
-                        key={timeSeriesGroup.id}
-                        locale={locale}
-                        timeSeriesGroup={timeSeriesGroup}
-                        selectedTime={selectedTime}
+            return (
+                <TimeSeriesChart
+                    key={timeSeriesGroup.id}
+                    locale={locale}
+                    timeSeriesGroup={timeSeriesGroup}
+                    selectedTime={selectedTime}
+                    selectedTimeRange={selectedTimeRange}
+                    dataTimeRange={dataTimeRange}
+                    selectTime={selectTime}
+                    selectTimeRange={selectTimeRange}
+                    removeTimeSeriesGroup={removeTimeSeriesGroup}
+                    completed={completed}
+                    showPointsOnly={showPointsOnly}
+                    showErrorBars={showErrorBars}
+                    placeInfos={placeInfos}
+                    places={places}
+                    selectPlace={selectPlace}
+                />
+            );
+        });
 
-                        selectedTimeRange={selectedTimeRange}
-                        dataTimeRange={dataTimeRange}
-                        selectTime={selectTime}
-                        selectTimeRange={selectTimeRange}
-                        removeTimeSeriesGroup={removeTimeSeriesGroup}
-                        completed={completed}
-                        showPointsOnly={showPointsOnly}
-                        showErrorBars={showErrorBars}
-                        placeInfos={placeInfos}
-                    />)
-            }
-        );
         if (charts.length === 0) {
             return null;
         }
+
         return (
             <div className={classes.chartContainer}>
                 <TimeRangeSlider